Split the StandardEditor (#50350)
